Refactored storage to factorize some code
